pyproject.toml
[tool.poetry]
name = "nfl-player-stats"
version = "0.1.0"
description = ""
authors = ["Kyle Clark <@kyleaclark>"]

[tool.poetry.dependencies]
python = "^3.7"
pendulum = "^2.0"
nflgame-redux = "=2.0.1a1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
